10082022 - added vendor registration flow by staff

+++ lib/components/searched_location_listview.dart
@@ -56,10 +56,10 @@ class SearchedLocationListViewState extends State<SearchedLocationListView> {
                 final geometry = detail.result.geometry!;
                 vendorManager.vendorRegistrationRequest.latitude = geometry.location.lat;
                 vendorManager.vendorRegistrationRequest.longitude = geometry.location.lng;
-                LatLng newLocation = LatLng(vendorManager.vendorRegistrationRequest.latitude,
-                    vendorManager.vendorRegistrationRequest.longitude);
+                LatLng newLocation = LatLng(vendorManager.vendorRegistrationRequest.latitude as double,
+                    vendorManager.vendorRegistrationRequest.longitude as double);
                 List<Placemark> placeMarks = await placemarkFromCoordinates(
-                    vendorManager.vendorRegistrationRequest.latitude, vendorManager.vendorRegistrationRequest.longitude);
+                    vendorManager.vendorRegistrationRequest.latitude as double, vendorManager.vendorRegistrationRequest.longitude as double);
                 locationManager.setCurrentLocation = '${placeMarks[0].locality}, ${placeMarks[0].subLocality}';
                 locationManager.mapController?.animateCamera(CameraUpdate.newCameraPosition(CameraPosition(target: newLocation, zoom: 14)));
                 locationManager.clearSearchedLocations();
